Split report sections into dedicated components
