+++ src/features/report/components/user/DetailedDescription.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function DetailedDescription({ onSubmit }) {
+    return <SectionForm sectionId="4" isLast onSubmit={onSubmit} />;
+}
