+++ src/features/report/components/user/HealthHistory.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function HealthHistory({ onSubmit }) {
+    return <SectionForm sectionId="6" isLast onSubmit={onSubmit} />;
+}
