+++ src/features/report/components/user/Impact.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function Impact({ onSubmit }) {
+    return <SectionForm sectionId="5" isLast onSubmit={onSubmit} />;
+}
