+++ src/features/report/components/user/InsuranceDetails.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function InsuranceDetails({ onSubmit }) {
+    return <SectionForm sectionId="2" isLast onSubmit={onSubmit} />;
+}
