+++ src/features/report/components/user/IntakeForm.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function IntakeForm({ onSubmit }) {
+    return <SectionForm sectionId="1" isLast onSubmit={onSubmit} />;
+}
