+++ src/features/report/components/user/PainEvaluation.jsx
@@ -1,0 +1,5 @@
+import SectionForm from './SectionForm';
+
+export default function PainEvaluation({ onSubmit }) {
+    return <SectionForm sectionId="3" isLast onSubmit={onSubmit} />;
+}
